Part 5.1 complete - Resource Conflict (add LoggingFilter and ExceptionMapping and fixed issues with visible log)

+++ src/main/java/com/mycompany/smartcampusapi/resources/SensorRoom.java
@@ -4,8 +4,8 @@
  */
 package com.mycompany.smartcampusapi.resources;
 
-import com.smartcampus.models.Room;
-import com.smartcampus.service.RoomService;
+import com.mycompany.smartcampus.models.Room;
+import com.mycompany.smartcampus.service.RoomService;
 
 /**
  *
@@ -53,6 +53,8 @@ public class SensorRoom {
     @Path("/{roomId}")
     @Produces(MediaType.APPLICATION_JSON)
     public Response getRoomId(@PathParam("roomId") String roomId){
+        
+        //removed switch statment for exception mapping
         Room room = roomService.getRoomById(roomId);
         
         if(room == null) {
@@ -68,23 +70,16 @@ public class SensorRoom {
     @Path("/{roomId}")
     @Produces(MediaType.APPLICATION_JSON)
     public Response deleteRoom(@PathParam("roomId") String roomId) {
-        String result = roomService.deleteRoom(roomId);
-        
-        switch(result){
-            case "SUCCESSFUL":
-                return Response.ok("{\"message\": \"Room successfully deleted\"}").build();
-            case "HAS_SENSORS":
-                //Specifc saftey logic for 2.2
-                return Response.status(Response.Status.CONFLICT)
-                        .entity("\"error\": \"Deletion blocked\", \"reason\": \"Room contains sensors.\"}")
-                        .build();
-            case "ROOM_NOT_FOUND":
-                return Response.status(Response.Status.NOT_FOUND)
-                        .entity("{\"error\": \"Room not found\"}")
-                        .build();
-            default:
-                return Response.status(Response.Status.INTERNAL_SERVER_ERROR).build();
+
+        Room room = roomService.getRoomById(roomId);
+        if(room == null){
+            return Response.status(Response.Status.NOT_FOUND)
+                    .entity("{\"error\": \"Room not found.\"}")
+                    .build();  
         }
+        
+        roomService.deleteRoom(roomId);
+        return Response.noContent().build();
     }
